Add a UserVoter to check if the current user can edit or delete a map, feest

+++ src/AppBundle/Controller/Admin/PhotoController.php
@@ -12,6 +12,7 @@ namespace AppBundle\Controller\Admin;
 use AppBundle\Entity\Photo;
 use AppBundle\Form\FolderForm;
 use AppBundle\Form\PhotoForm;
+use AppBundle\Security\UserVoter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
@@ -64,7 +65,11 @@ class PhotoController extends Controller
      */
     public function EditPhotoAction(Request $request, Photo $photo)
     {
-//        $photo = new Photo();
+
+//      use the UserVoter to check if this is the creator of the foto
+        $this->denyAccessUnlessGranted(UserVoter::EDIT, $photo);
+
+//
         $form = $this->createForm(PhotoForm::class, $photo);
         $form->handleRequest($request);
 
@@ -98,6 +103,8 @@ class PhotoController extends Controller
      */
     public function deletePhotoAction(Request $request, Photo $photo)
     {
+        //      use the UserVoter to check if this is the creator of the foto
+        $this->denyAccessUnlessGranted(UserVoter::DELETE, $photo);
 
         $em = $this->getDoctrine()->getManager();
         $em->remove($photo);
